+++ src/Symfony/Bridge/Doctrine/DataCollector/DoctrineDataCollector.php
@@ -71,11 +71,17 @@ class DoctrineDataCollector extends DataCollector
         $this->debugDataHolder->reset();
     }
 
+    /**
+     * @return array
+     */
     public function getManagers()
     {
         return $this->data['managers'];
     }
 
+    /**
+     * @return array
+     */
     public function getConnections()
     {
         return $this->data['connections'];
@@ -89,11 +95,17 @@ class DoctrineDataCollector extends DataCollector
         return array_sum(array_map('count', $this->data['queries']));
     }
 
+    /**
+     * @return array
+     */
     public function getQueries()
     {
         return $this->data['queries'];
     }
 
+    /**
+     * @return int
+     */
     public function getTime()
     {
         $time = 0;