+++ src/Symfony/Component/DependencyInjection/Tests/Compiler/RegisterEnvVarProcessorsPassTest.php
@@ -50,6 +50,7 @@ class RegisterEnvVarProcessorsPassTest extends TestCase
             'require' => ['bool', 'int', 'float', 'string', 'array'],
             'enum' => [\BackedEnum::class],
             'shuffle' => ['array'],
+            'defined' => ['bool'],
         ];
 
         $this->assertSame($expected, $container->getParameterBag()->getProvidedTypes());