+++ src/Symfony/Component/HttpKernel/DataCollector/LoggerDataCollector.php
@@ -16,6 +16,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Log\DebugLoggerInterface;
+use Symfony\Component\VarDumper\Cloner\Data;
 
 /**
  * @author Fabien Potencier <fabien@symfony.com>
@@ -67,12 +68,12 @@ class LoggerDataCollector extends DataCollector implements LateDataCollectorInte
         $this->currentRequest = null;
     }
 
-    public function getLogs()
+    public function getLogs(): Data|array
     {
         return $this->data['logs'] ?? [];
     }
 
-    public function getProcessedLogs()
+    public function getProcessedLogs(): array
     {
         if (null !== $this->processedLogs) {
             return $this->processedLogs;
@@ -115,7 +116,7 @@ class LoggerDataCollector extends DataCollector implements LateDataCollectorInte
         return $this->processedLogs = $logs;
     }
 
-    public function getFilters()
+    public function getFilters(): array
     {
         $filters = [
             'channel' => [],
@@ -146,32 +147,32 @@ class LoggerDataCollector extends DataCollector implements LateDataCollectorInte
         return $filters;
     }
 
-    public function getPriorities()
+    public function getPriorities(): Data|array
     {
         return $this->data['priorities'] ?? [];
     }
 
-    public function countErrors()
+    public function countErrors(): int
     {
         return $this->data['error_count'] ?? 0;
     }
 
-    public function countDeprecations()
+    public function countDeprecations(): int
     {
         return $this->data['deprecation_count'] ?? 0;
     }
 
-    public function countWarnings()
+    public function countWarnings(): int
     {
         return $this->data['warning_count'] ?? 0;
     }
 
-    public function countScreams()
+    public function countScreams(): int
     {
         return $this->data['scream_count'] ?? 0;
     }
 
-    public function getCompilerLogs()
+    public function getCompilerLogs(): Data
     {
         return $this->cloneVar($this->getContainerCompilerLogs($this->data['compiler_logs_filepath'] ?? null));
     }